refactor(forms): consolidate insider-trading value helpers into shared module + Form_C whitespace name fallback
Code-review follow-ups on the cherry-picked #118/#157 fixes:

- Remove the duplicate src/sec/forms/insider-trading/_valueHelpers.ts (and
  its test) reintroduced by the #118 cherry-pick. main already carries the
  shared src/sec/forms/_valueHelpers.ts (imported by Form_C/1-A/1-Z/1-K),
  whose own note asked for the duplicate to be dropped once #118 landed.
  Form_144 and OwnershipDocument now import from ../_valueHelpers.
- Teach the shared strWrapped to accept a bare string so OwnershipDocument's
  untagged leaves (documentType, issuerName, ...) keep their old behaviour;
  this is additive (no prior production consumer of the shared *Wrapped).
  Ported the bare-string / no-.value edge-case coverage into the shared test.
- Form_C: strScalar(issuer.nameOfIssuer) || existing?.name || "" — a
  whitespace-only nameOfIssuer now falls back to the existing row's name
  (matching the empty-string case and the helper trim philosophy), and the
  dead third operand (a || b || a) is dropped.
