+++ src/sec/forms/insider-trading/Form_144.storage.ts
@@ -31,7 +31,7 @@ import { getActiveSlot } from "../../../storage/versioning/getActiveSlot";
 import { formToExtractorId } from "../../../storage/versioning/extractorIds";
 import { Form144Repo } from "../../../storage/form144/Form144Repo";
 import type { Form144 } from "./Form_144.schema";
-import { numScalar as num, strScalar as str } from "./_valueHelpers";
+import { numScalar as num, strScalar as str } from "../_valueHelpers";
 
 type AddressShape = NonNullable<
   NonNullable<Form144["formData"]>["issuerInfo"]
